requirements.txt
httpx>=0.24.0
aiofiles>=23.1.0
lxml>=4.9.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
pyyaml>=6.0
tqdm>=4.65.0
urllib3>=2.0.0
aiohttp>=3.8.0
pytest>=7.4.0
